requirements.txt
tensorflow==2.17.0
gradio
pillow
numpy
matplotlib
